refactor core out into separate modules
