+++ objex/__init__.py
@@ -1,1 +1,2 @@
-from .core import dump_graph, make_analysis_db, Reader, CLI
+from .exporter import dump_graph
+from .explorer import make_analysis_db, Reader, Console
