+++ objex/__main__.py
@@ -1,9 +1,9 @@
 import sys
 
-from . import core
+from . import explorer
 
 try:
-    core.CLI(core.Reader(sys.argv[-1])).run()
+    explorer.Console(explorer.Reader(sys.argv[-1])).run()
 except:
     import traceback; traceback.print_exc()
     import pdb; pdb.post_mortem()
